full dulex and basic transfer state display
实现全双工文件传输；实现基本的传输状态显示

+++ crc_verify.h
@@ -6,8 +6,8 @@
 #include "stdint.h"
 #include "stdio.h"
 
-uint16_t crc16_CCITT(uint8_t* datas,uint8_t len);
-uint16_t crc16_CCITT(char* data,uint8_t len);
-bool check_CCITT(char* data, uint8_t len);
+uint16_t crc16_CCITT(uint8_t* datas,uint16_t len);
+uint16_t crc16_CCITT(char* data,uint16_t len);
+bool check_CCITT(char* data, uint16_t len);
 
 #endif // CRC_VERIFY_H
